Add metrics for ratin_service queue length.
Quick typo cleanup

Adapt tests after rebase
